Fix: merge userMenu login
