+++ src/components/kakaoLogin/KakaoLogin.tsx
@@ -53,7 +53,7 @@ export default function KakaoLogin({ KakaoLocation }: KakaoLoginProps) {
   return (
     <div className={`${KakaoLocation === '카카오 로그인' ? 'w-auto' : 'w-full'}`}>
       {KakaoLocation === '카카오 로그인' ? (
-        <Button className='text-black text-2xl font-bold w-full' onClick={handleLoginClick}>
+        <Button className='text-black text-sm font-bold w-full' onClick={handleLoginClick}>
           카카오로 로그인하기
         </Button>
       ) : (
