+++ src/components/navigation/navibarButtons/styles.module.css
@@ -10,6 +10,7 @@
 .hrSect::before,
 .hrSect::after {
   content: '';
+  font-size: 14px;
   flex-grow: 2;
   background: rgba(0, 0, 0, 0.35);
   height: 1px;
